fix(mastracode): resolve models through gateway class
Route MastraCode model ID resolution through MastraCodeGateway directly so auth and provider selection stay inside the gateway implementation while Harness accepts the resulting provider model configs.
