+++ mastracode/src/agents/__tests__/model.test.ts
@@ -92,6 +92,8 @@ vi.mock('ai', () => ({
 
 // Mock ModelRouterLanguageModel and MastraGateway
 vi.mock('@mastra/core/llm', () => ({
+  MastraModelGateway: class MockMastraModelGateway {},
+  PROVIDER_REGISTRY: {},
   ModelRouterLanguageModel: vi.fn(function (
     this: Record<string, unknown>,
     config: string | { id: string; url?: string; apiKey?: string; headers?: Record<string, string> },
@@ -171,6 +173,7 @@ import {
   getDynamicModel,
   getAnthropicApiKey,
   getOpenAIApiKey,
+  MastraCodeGateway,
   resolveAuth,
 } from '../model.js';
 
@@ -200,6 +203,7 @@ describe('resolveModel', () => {
     delete process.env.OPENAI_API_KEY;
     delete process.env.OPENAI_BASE_URL;
     delete process.env.MOONSHOT_AI_API_KEY;
+    delete process.env.MASTRA_GATEWAY_API_KEY;
     delete process.env.MASTRA_GATEWAY_URL;
   });
 
@@ -225,6 +229,9 @@ describe('resolveModel', () => {
       mastraGatewayBaseUrl: 'https://gateway-api.mastra.ai',
       routeThroughMastraGateway: false,
     });
+
+    expect(gateway).toBeInstanceOf(MastraCodeGateway);
+    expect(gateway.id).toBe('mastracode');
 
     await expect(gateway.fetchProviders()).resolves.toMatchObject({
       'acme-models': {
@@ -254,6 +261,36 @@ describe('resolveModel', () => {
       resolveModel('anthropic/claude-sonnet-4-20250514');
 
       expect(opencodeClaudeMaxProvider).toHaveBeenCalledWith('claude-sonnet-4-20250514', { headers: undefined });
+    });
+
+    it('parses provider/model ids and delegates directly through the MastraCode gateway', () => {
+      const resolveAuthSpy = vi.spyOn(MastraCodeGateway.prototype, 'resolveAuth');
+      const resolveLanguageModelSpy = vi.spyOn(MastraCodeGateway.prototype, 'resolveLanguageModel');
+
+      try {
+        const result = resolveModel('anthropic/claude-sonnet-4-20250514') as Record<string, unknown>;
+
+        expect(result.__provider).toBe('claude-max-oauth');
+        expect(resolveAuthSpy).toHaveBeenCalledWith({
+          gatewayId: 'mastracode',
+          providerId: 'anthropic',
+          modelId: 'claude-sonnet-4-20250514',
+          routerId: 'mastracode/anthropic/claude-sonnet-4-20250514',
+        });
+        expect(resolveLanguageModelSpy).toHaveBeenCalledWith({
+          providerId: 'anthropic',
+          modelId: 'claude-sonnet-4-20250514',
+          apiKey: '',
+          headers: undefined,
+        });
+        expect(ModelRouterLanguageModel).not.toHaveBeenCalledWith(
+          { id: 'mastracode/anthropic/claude-sonnet-4-20250514', headers: undefined },
+          expect.any(Array),
+        );
+      } finally {
+        resolveAuthSpy.mockRestore();
+        resolveLanguageModelSpy.mockRestore();
+      }
     });
 
     it('uses API key when stored credential is api_key, even if isLoggedIn reports true', () => {
@@ -510,21 +547,7 @@ describe('resolveModel', () => {
       expect(MastraGateway).toHaveBeenCalledWith({
         baseUrl: 'https://gateway-api.mastra.ai',
       });
-      expect(ModelRouterLanguageModel).toHaveBeenCalledWith(
-        { id: 'mastracode/anthropic/claude-sonnet-4', headers: undefined },
-        [expect.objectContaining({ id: 'mastracode', resolveAuth: expect.any(Function) })],
-      );
-      const gateway = (
-        result.customGateways as Array<{
-          resolveAuth: (request: Record<string, string>) => Record<string, string> | undefined;
-        }>
-      )[0];
-      expect(
-        gateway?.resolveAuth({ gatewayId: 'mastracode', providerId: 'anthropic', modelId: 'claude-sonnet-4' }),
-      ).toEqual({
-        apiKey: 'msk_gateway_key_123',
-        source: 'gateway',
-      });
+      expect(ModelRouterLanguageModel).not.toHaveBeenCalled();
     });
 
     it('routes explicit mastra-prefixed anthropic OAuth model through custom gateway middleware', () => {
@@ -540,27 +563,7 @@ describe('resolveModel', () => {
       expect(result.__provider).toBe('anthropic-direct');
       expect(result.__wrapped).toBe(true);
       expect(MastraGateway).toHaveBeenCalledWith({ baseUrl: 'https://gateway-api.mastra.ai' });
-      expect(ModelRouterLanguageModel).toHaveBeenCalledWith(
-        { id: 'mastracode/anthropic/claude-sonnet-4', headers: undefined },
-        [
-          expect.objectContaining({
-            id: 'mastracode',
-            resolveAuth: expect.any(Function),
-            resolveLanguageModel: expect.any(Function),
-          }),
-        ],
-      );
-
-      const gateway = (
-        result.customGateways as Array<{
-          resolveLanguageModel: (args: Record<string, unknown>) => Record<string, unknown>;
-        }>
-      )[0];
-      const resolved = gateway?.resolveLanguageModel({
-        providerId: 'anthropic',
-        modelId: 'claude-sonnet-4',
-        apiKey: 'msk_gateway_key_123',
-      });
+      expect(ModelRouterLanguageModel).not.toHaveBeenCalled();
 
       expect(createAnthropic).toHaveBeenCalledWith(
         expect.objectContaining({
@@ -574,9 +577,9 @@ describe('resolveModel', () => {
         'Bearer msk_gateway_key_123',
       );
       expect(wrapLanguageModel).toHaveBeenCalled();
-      expect(resolved?.__wrapped).toBe(true);
-      expect(resolved?.__provider).toBe('anthropic-direct');
-      expect(resolved?.modelId).toBe('claude-sonnet-4');
+      expect(result.__wrapped).toBe(true);
+      expect(result.__provider).toBe('anthropic-direct');
+      expect(result.modelId).toBe('claude-sonnet-4');
       expect(opencodeClaudeMaxProvider).not.toHaveBeenCalled();
     });
 
@@ -589,22 +592,10 @@ describe('resolveModel', () => {
       });
 
       const result = resolveModel('mastra/anthropic/claude-opus-4.6') as Record<string, unknown>;
-      const gateway = (
-        result.customGateways as Array<{
-          resolveLanguageModel: (args: Record<string, unknown>) => Record<string, unknown>;
-        }>
-      )[0];
-      const resolved = gateway?.resolveLanguageModel({
-        providerId: 'anthropic',
-        modelId: 'claude-opus-4.6',
-        apiKey: 'msk_gateway_key_123',
-      });
 
       expect(result.__provider).toBe('anthropic-direct');
       expect(result.__wrapped).toBe(true);
-      expect(resolved?.__provider).toBe('anthropic-direct');
-      expect(resolved?.__wrapped).toBe(true);
-      expect(resolved?.modelId).toBe('claude-opus-4-6');
+      expect(result.modelId).toBe('claude-opus-4-6');
     });
 
     it('routes explicit mastra-prefixed openai OAuth model through custom gateway Codex middleware', () => {
@@ -620,24 +611,7 @@ describe('resolveModel', () => {
       expect(result.__provider).toBe('openai-direct');
       expect(result.__wrapped).toBe(true);
       expect(MastraGateway).toHaveBeenCalledWith({ baseUrl: 'https://gateway-api.mastra.ai' });
-      expect(ModelRouterLanguageModel).toHaveBeenCalledWith({ id: 'mastracode/openai/gpt-4o', headers: undefined }, [
-        expect.objectContaining({
-          id: 'mastracode',
-          resolveAuth: expect.any(Function),
-          resolveLanguageModel: expect.any(Function),
-        }),
-      ]);
-
-      const gateway = (
-        result.customGateways as Array<{
-          resolveLanguageModel: (args: Record<string, unknown>) => Record<string, unknown>;
-        }>
-      )[0];
-      const resolved = gateway?.resolveLanguageModel({
-        providerId: 'openai',
-        modelId: 'gpt-4o',
-        apiKey: 'msk_gateway_key_123',
-      });
+      expect(ModelRouterLanguageModel).not.toHaveBeenCalled();
 
       expect(createOpenAI).toHaveBeenCalledWith(
         expect.objectContaining({
@@ -655,9 +629,9 @@ describe('resolveModel', () => {
         rewriteUrl: false,
       });
       expect(wrapLanguageModel).toHaveBeenCalled();
-      expect(resolved?.__wrapped).toBe(true);
-      expect(resolved?.__provider).toBe('openai-direct');
-      expect(resolved?.modelId).toBe('gpt-4o');
+      expect(result.__wrapped).toBe(true);
+      expect(result.__provider).toBe('openai-direct');
+      expect(result.modelId).toBe('gpt-4o');
       expect(openaiCodexProvider).not.toHaveBeenCalled();
     });
 
@@ -692,18 +666,6 @@ describe('resolveModel', () => {
       expect(MastraGateway).toHaveBeenCalledWith({
         baseUrl: 'https://gateway-api.mastra.ai',
       });
-
-      const gateway = (
-        result.customGateways as Array<{
-          resolveLanguageModel: (args: Record<string, unknown>) => Record<string, unknown>;
-        }>
-      )[0];
-      const resolved = gateway?.resolveLanguageModel({
-        providerId: 'google',
-        modelId: 'gemini-2.0-flash',
-        apiKey: 'msk_gateway_key_123',
-      });
-      expect(resolved?.__provider).toBe('mastra-gateway-delegate');
     });
 
     it('custom provider bypasses gateway', () => {
@@ -748,20 +710,18 @@ describe('resolveModel', () => {
       });
     });
 
-    it('passes harness headers to ModelRouterLanguageModel for explicit mastra-prefixed models', () => {
+    it('passes harness headers to the gateway-resolved model', () => {
       mockAuthStorageInstance.get.mockReturnValue(undefined);
 
-      resolveModel('mastra/anthropic/claude-sonnet-4', {
+      const result = resolveModel('mastra/anthropic/claude-sonnet-4', {
         requestContext: makeRequestContext({ threadId: 'thread-123', resourceId: 'resource-456' }),
-      });
+      }) as Record<string, unknown>;
 
-      expect(ModelRouterLanguageModel).toHaveBeenCalledWith(
-        {
-          id: 'mastracode/anthropic/claude-sonnet-4',
-          headers: { 'x-thread-id': 'thread-123', 'x-resource-id': 'resource-456' },
-        },
-        expect.any(Array),
-      );
+      expect((result.args as Record<string, unknown>).headers).toEqual({
+        'x-thread-id': 'thread-123',
+        'x-resource-id': 'resource-456',
+      });
+      expect(ModelRouterLanguageModel).not.toHaveBeenCalled();
     });
 
     it('skips gateway when no API key is stored and no env var', () => {
